+++ core-bundle/src/Resources/contao/library/Contao/Template.php
@@ -527,16 +527,19 @@ abstract class Template extends Controller
 	/**
 	 * Generate the markup for a JavaScript tag
 	 *
-	 * @param string  $src   The script path
-	 * @param boolean $async True to add the async attribute
-	 * @param mixed   $mtime The file mtime
+	 * @param string      $src   The script path
+	 * @param boolean     $async True to add the async attribute
+	 * @param mixed       $mtime The file mtime
+	 * @param string|null $hash  An optional integrity hash
 	 *
 	 * @return string The markup string
 	 */
-	public static function generateScriptTag($src, $async=false, $mtime=false)
+	public static function generateScriptTag($src, $async=false, $mtime=false, $hash=null)
 	{
+		$external = preg_match('@^https?://@', $src);
+
 		// Add the filemtime if not given and not an external file
-		if ($mtime === null && !preg_match('@^https?://@', $src))
+		if ($mtime === null && !$external)
 		{
 			$container = \System::getContainer();
 			$rootDir = $container->getParameter('kernel.project_dir');
@@ -562,7 +565,7 @@ abstract class Template extends Controller
 			$src .= '?v=' . substr(md5($mtime), 0, 8);
 		}
 
-		return '<script src="' . $src . '"' . ($async ? ' async' : '') . '></script>';
+		return '<script src="' . $src . '"' . ($async ? ' async' : '') . ($hash ? ' integrity="' . $hash . '"' : '') . ($external ? ' crossorigin="anonymous"' : '') . '></script>';
 	}
 
 	/**