+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## DEV
 
+ * Add the integrity attribute when loading jQuery from CDN (see contao/core-bundle#702).
  * Add methods to retrive past and upcoming dates in the event reader (see #175).
  * Associate comments with members (see contao/comments-bundle#7).
  * Add content disposition to download elements (see #20).